Move all the property binding to DataContext
Put this all in a seperate file so it's easier to look at things together
